+++ navigator-example/src/main/java/com/zhuinden/navigatorexample/SecondKey.java
@@ -1,6 +1,7 @@
 package com.zhuinden.navigatorexample;
 
 import android.os.Parcelable;
+import android.support.annotation.NonNull;
 
 import com.google.auto.value.AutoValue;
 import com.zhuinden.navigator.ViewChangeHandler;
@@ -23,12 +24,13 @@ public abstract class SecondKey
     }
 
     @Override
-    public ViewChangeHandler getViewChangeHandler() {
+    @NonNull
+    public ViewChangeHandler viewChangeHandler() {
         return new SegueViewChangeHandler();
     }
 
     @Override
-    public String getTitle() {
+    public String title() {
         return "Second";
     }
 }