+++ navigator/src/main/java/com/zhuinden/navigator/Navigator.java
@@ -34,13 +34,30 @@ import com.zhuinden.simplestack.StateChanger;
 
 import java.util.List;
 
+/**
+ * Convenience class to hide lifecycle integration using retained fragment.
+ * Essentially, a replacement for BackstackDelegate.
+ *
+ * It can be either configured via {@link Navigator#configure()}, or installed with default settings using {@link Navigator#install(Activity, ViewGroup, List)}.
+ */
 public class Navigator {
+    /**
+     * A configurer for {@link Navigator}.
+     */
     public static class Installer {
-        StateChanger stateChanger = new InternalStateChanger.NoOpStateChanger();
+        StateChanger stateChanger;
         BackstackManager.StateClearStrategy stateClearStrategy = new DefaultStateClearStrategy();
         KeyParceler keyParceler = new DefaultKeyParceler();
         boolean isInitializeDeferred = false;
 
+        /**
+         * Sets the state changer used by the navigator's backstack.
+         *
+         * If not set, then {@link DefaultStateChanger} is used, which requires keys to be {@link StateKey}.
+         *
+         * @param stateChanger if set, cannot be null.
+         * @return the installer
+         */
         public Installer setStateChanger(@NonNull StateChanger stateChanger) {
             if(stateChanger == null) {
                 throw new IllegalArgumentException("If set, StateChanger cannot be null!");
@@ -49,6 +66,12 @@ public class Navigator {
             return this;
         }
 
+        /**
+         * Sets the key parceler for parcelling state keys.
+         *
+         * @param keyParceler cannot be null if set
+         * @return the installer
+         */
         public Installer setKeyParceler(@NonNull KeyParceler keyParceler) {
             if(keyParceler == null) {
                 throw new IllegalArgumentException("If set, KeyParceler cannot be null!");
@@ -57,6 +80,12 @@ public class Navigator {
             return this;
         }
 
+        /**
+         * Sets the state clear strategy used to clear the stored state in BackstackManager after there are no queued state changes left.
+         *
+         * @param stateClearStrategy if set, it cannot be null
+         * @return the installer
+         */
         public Installer setStateClearStrategy(@NonNull BackstackManager.StateClearStrategy stateClearStrategy) {
             if(stateClearStrategy == null) {
                 throw new IllegalArgumentException("If set, StateClearStrategy cannot be null!");
@@ -65,12 +94,30 @@ public class Navigator {
             return this;
         }
 
+        /**
+         * Sets if after initialization, the state changer should only be set when {@link Navigator#executeDeferredInitialization(Context)} is called.
+         * Typically needed to setup the backstack for dependency injection module.
+         *
+         * @param isInitializeDeferred if call to executing deferred initialization is needed
+         * @return the installer
+         */
         public Installer setDeferredInitialization(boolean isInitializeDeferred) {
             this.isInitializeDeferred = isInitializeDeferred;
             return this;
         }
 
+        /**
+         * Installs the {@link BackstackHost}.
+         *
+         * @param activity    the activity
+         * @param container   the container
+         * @param initialKeys the initial keys.
+         * @return
+         */
         public Backstack install(@NonNull Activity activity, @NonNull ViewGroup container, @NonNull List<Object> initialKeys) {
+            if(stateChanger == null) {
+                stateChanger = new DefaultStateChanger(activity, container);
+            }
             return Navigator.install(this, activity, container, initialKeys);
         }
     }
@@ -78,10 +125,24 @@ public class Navigator {
     private Navigator() {
     }
 
+    /**
+     * Creates an {@link Installer} to configure the {@link Navigator}.
+     *
+     * @return the installer
+     */
     public static Installer configure() {
         return new Installer();
     }
 
+    /**
+     * Installs the {@link Navigator} with default parameters.
+     *
+     * This means that {@link DefaultStateChanger} and {@link DefaultStateClearStrategy} are used.
+     *
+     * @param activity the activity which will host the backstack
+     * @param container the container in which custom viewgroups are hosted (to save its child's state in onSaveInstanceState())
+     * @param initialKeys the keys used to initialize the backstack
+     */
     public static void install(@NonNull Activity activity, @NonNull ViewGroup container, @NonNull List<Object> initialKeys) {
         install(configure(), activity, container, initialKeys);
     }
@@ -102,7 +163,7 @@ public class Navigator {
             activity.getFragmentManager().beginTransaction().add(backstackHost, "NAVIGATOR_BACKSTACK_HOST").commit();
             activity.getFragmentManager().executePendingTransactions();
         }
-        backstackHost.externalStateChanger = installer.stateChanger;
+        backstackHost.stateChanger = installer.stateChanger;
         backstackHost.keyParceler = installer.keyParceler;
         backstackHost.stateClearStrategy = installer.stateClearStrategy;
         backstackHost.initialKeys = initialKeys;
@@ -110,20 +171,43 @@ public class Navigator {
         return backstackHost.initialize(installer.isInitializeDeferred);
     }
 
-    public static void executeDeferredInitialization(Activity activity) {
+    /**
+     * If {@link Installer#setDeferredInitialization(boolean)} was set to true, then this will initialize the backstack using the state changer.
+     *
+     * @param context the context to which an activity belongs that hosts the backstack
+     */
+    public static void executeDeferredInitialization(Context context) {
+        Activity activity = findActivity(context);
         BackstackHost backstackHost = findBackstackHost(activity);
         backstackHost.initialize(false);
     }
 
+    /**
+     * Gets the backstack that belongs to the Activity which hosts the backstack.
+     *
+     * @param context the context
+     * @return the backstack
+     */
     public static Backstack getBackstack(Context context) {
         BackstackHost backstackHost = getBackstackHost(context);
         return backstackHost.getBackstack();
     }
 
-    public static boolean onBackPressed(Activity activity) {
-        return getBackstack(activity).goBack();
+    /**
+     * Delegates back press call to the backstack of the navigator.
+     *
+     * @param context the Context that belongs to an Activity which hosts the backstack.
+     * @return true if a state change was handled or is in progress, false otherwise
+     */
+    public static boolean onBackPressed(Context context) {
+        return getBackstack(context).goBack();
     }
 
+    /**
+     * Persists the view hierarchy state and optional StateBundle.
+     *
+     * @param view the view (can be Bundleable)
+     */
     public static void persistViewToState(@Nullable View view) {
         if(view != null) {
             Context context = view.getContext();
@@ -132,6 +216,11 @@ public class Navigator {
         }
     }
 
+    /**
+     * Restores the view hierarchy state and optional StateBundle.
+     *
+     * @param view the view (can be Bundleable)
+     */
     public static void restoreViewFromState(@NonNull View view) {
         if(view == null) {
             throw new NullPointerException("You cannot restore state into null view!");
@@ -141,15 +230,22 @@ public class Navigator {
         backstackHost.backstackManager.restoreViewFromState(view);
     }
 
-    public static SavedState getSavedState(@NonNull Context context, @NonNull StateKey stateKey) {
+    /**
+     * Get the saved state for a given key.
+     *
+     * @param context the context to which an Activity belongs that hosts a backstack
+     * @param key     the key
+     * @return the saved state
+     */
+    public static SavedState getSavedState(@NonNull Context context, @NonNull Object key) {
         if(context == null) {
             throw new NullPointerException("context cannot be null");
         }
-        if(stateKey == null) {
-            throw new NullPointerException("stateKey cannot be null");
+        if(key == null) {
+            throw new NullPointerException("key cannot be null");
         }
         BackstackHost backstackHost = getBackstackHost(context);
-        return backstackHost.backstackManager.getSavedState(stateKey);
+        return backstackHost.backstackManager.getSavedState(key);
     }
 
     private static BackstackHost findBackstackHost(Activity activity) {